Detect and fix user pasting JSON into SQL field
This builds on the previous commit [1] and should make it nearly
impossible for a new user to make a mistake when pasting their JSON into
the wrong textarea.

The same feature could be implemented for pasting the SQL into the JSON
input box, but detecting SQL is a bit harder, and the user is already
getting a JSON parsing error in this case, so it seems less important
for now.

[1] 1151672cf8f52bd80c908adb9c2cb19c2b08250e